driver: bma400 setup and test with interrupt
BMA400 APIs still needs to be refactored. There are still some bugs such
as the xyz readings fluctuating greatly even though they move
in the right direction as the sensor is tilted. I tried to follow the
datasheet closely for formulas before but they didn't calculate right.
The solution I have so far is a hack and is temporary.