test: enable uart flow control
Cannot verify yet if the flow control is actually working because
the ble module communicates fine with the mcu even if the CTS and RTS pins
are disconnected. Need to investigate here more.

Also this is an attempt to fix the issue of getting stuck in uart transmit if the
payload is over 180 bytes.
Now, we can go up to 252 (the max is 256. However, there is a 4 byte header).
It looks like the 4 byte header is part of the 'payload' or counted
as part of the MTU.
Note: tests show that the MCU is transmitting to the ble module. But
the BLE Scanner app is not picking up on those transmissions. Still need to
confirm this. Perhaps the app itself has some payload limitations to receive.